+++ server/src/routes/v1/routesV1.ts
@@ -1,5 +1,4 @@
 import * as express from "express";
-import { login } from "./login/login";
 import { session } from "./session/session";
 import { booksRouter } from "./books/booksRouter";
 import { jwtTokenVerification } from "../../middlewares/jwt";
@@ -8,13 +7,12 @@ import { typesRouter } from "./types/typesRouter";
 import { languagesRouter } from "./languages/languagesRouter";
 import { countryRouter } from "./country_origin/countryRouter";
 import { data_saver } from "./data_saver/data_saver";
-import { createAccount } from "./login/createAccount";
+import { accountRouter } from "./account/acountRouter";
 
 export const v1Routes = express.Router();
 
 /** Session Routers */
-v1Routes.post("/login", login);
-v1Routes.post("/createaccount", createAccount);
+v1Routes.use("/account", accountRouter);
 v1Routes.get("/session", session);
 
 /** Open API access Routers */